refactor(api): change typeorm to mongoose

+++ apps/api/src/features/profiles/profiles.controller.ts
@@ -6,40 +6,55 @@ import {
   Patch,
   Param,
   Delete,
-  UseGuards, Request,
+  UseGuards,
+  Request,
+  ClassSerializerInterceptor,
+  UseInterceptors,
+  HttpCode,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { ProfilesService } from './profiles.service';
-import { CreateProfileDto } from './dto/create-profile.dto';
-import { UpdateProfileDto } from './dto/update-profile.dto';
+import { CreateProfileDto } from './models/create-profile.dto';
+import { UpdateProfileDto } from './models/update-profile.dto';
 import { JwtAuthGuard } from '../auth/jwt-auth.guard';
+import { JwtIsAdminAuthGuard } from '../auth/jwt-is-admin-auth.guard';
 
 @Controller('profiles')
 export class ProfilesController {
   constructor(private readonly profilesService: ProfilesService) {}
 
-  @UseGuards(JwtAuthGuard)
   @Post()
-  create(@Body() createProfileDto: CreateProfileDto, @Request() req) {
-    return this.profilesService.create(createProfileDto);
+  @HttpCode(201)
+  @UseGuards(JwtAuthGuard)
+  async create(@Request() req, @Body() createProfileDto: CreateProfileDto) {
+    await this.profilesService.create(req.user.id, createProfileDto);
   }
 
   @Get()
+  @UseGuards(JwtIsAdminAuthGuard)
   findAll() {
     return this.profilesService.findAll();
   }
 
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.profilesService.findOne(id);
+  @Get(':fullname')
+  @UseInterceptors(ClassSerializerInterceptor)
+  findOne(@Param('fullname') fullname: string) {
+    return this.profilesService.findOne(fullname);
   }
 
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateProfileDto: UpdateProfileDto) {
-    return this.profilesService.update(id, updateProfileDto);
+  @Patch()
+  @UseGuards(JwtAuthGuard)
+  update(@Request() req, @Body() updateProfileDto: UpdateProfileDto) {
+    return this.profilesService.update(req.user.id, updateProfileDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.profilesService.remove(id);
+  @UseGuards(JwtAuthGuard)
+  async remove(@Request() req, @Param('id') id: string) {
+    if (req.user.username === 'admin' || req.user.id == id) {
+      await this.profilesService.remove(id);
+      return;
+    }
+    throw new UnauthorizedException();
   }
 }
